requirements.txt
# Base system
Django==3.2.7
pytz==2020.1
dj-database-url==0.4.1

# Parsing
lxml==4.6.3

# API
djangorestframework==3.12.4
django-filter==2.4.0
djangorestframework-camel-case==1.2.0
Markdown==3.3.4

#XMPP
sleekxmpp==1.3.3
dnspython==2.1.0
